workload/schemachange: disable random load test
Previously, the random load test was enabled on 21.2 and
hitting random failures linked to issues inside the test
itself. This led to issues being opened daily on nightly
builds. To address this, this patch will disable the
schema change random load test on 21.2 where the test
itself is unstable.

Release note: None