catalog/lease: handle transition to multiregion system database
Previously, when converting the system database to multi-region we could
run into transient errors during the process from the wait for one
version logic. This would happen because the lease descriptor used by
the transaction could be non-MR while the cached system database
descriptor could be MR. This is expected since the two can be from
different timestamps. To help address this, this patch will retry when
InvalidFunction and InvalidParameterValue pgcodes with references to
crdb_internal_region types are observed, which will happen if the lease
table has not update the crdb_region columns type. The retry attempt
will intentionally reference the region type as byte.

Fixes: #138335

Release note: None