catalog/lease: optimize bulk lease API for warm case
Previously, there was a lot of overhead in terms of mutex acquistions
when a bulk prefetch was request. This was because we would lock each
descriptor object to check if its populated. This patch simplifies the
logic to only check if the entry exists, and does more expensive
processing for missing entries. It also optimizes the warm case by
avoiding re-resolving descriptor states.

Release note: None